Updated dependencies
