--- FrontEnd/.travis.yml
@@ -1,20 +1,0 @@
-dist: trusty
-sudo: required
-language: node_js
-node_js:
-  - "5"
-os:
-  - linux
-env:
-  global:
-    - DBUS_SESSION_BUS_ADDRESS=/dev/null
-    - DISPLAY=:99.0
-    - CHROME_BIN=chromium-browser
-before_script:
-  - sh -e /etc/init.d/xvfb start
-install:
-  - npm install
-script:
-  - npm run lint
-  - npm run test:once
-  - npm run e2e
