--- FrontEnd/bs-config.e2e.json
@@ -1,14 +1,0 @@
-{
-  "open": false,
-  "logLevel": "silent",
-  "port": 8080,
-  "server": {
-    "baseDir": "src",
-    "routes": {
-      "/node_modules": "node_modules"
-    },
-    "middleware": {
-      "0": null
-    }
-  }
-}
