--- FrontEnd/bs-config.json
@@ -1,8 +1,0 @@
-{
-  "server": {
-    "baseDir": "src",
-    "routes": {
-      "/node_modules": "node_modules"
-    }
-  }
-}
