--- FrontEnd/non-essential-files.txt
@@ -1,15 +1,0 @@
-.git
-.gitignore
-.travis.yml
-*.spec*.ts
-bs-config.e2e.json
-CHANGELOG.md
-e2e
-favicon.ico
-karma.conf.js
-karma-test-shim.js
-LICENSE
-non-essential-files.txt
-non-essential-files.osx.txt
-protractor.config.js
-README.md
