+++ FrontEnd/src/app/app.module.ts
@@ -1,40 +1,14 @@
-import { NgModule }      from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import {ReactiveFormsModule,FormsModule} from '@angular/forms';
-import { AppComponent }  from './app.component';
-import {BorrowerComponent} from './components/BorrowerComponent';
-import {SearchComponent} from './components/SearchComponent';
-import { RouterModule, Routes } from '@angular/router';
-import { HttpModule, JsonpModule } from '@angular/http';
-import {CheckoutBook} from "./components/CheckOutBook";
-
-import {CheckInBook} from "./components/CheckInBook";
-import {SearchCheckInBook} from "./components/SearchCheckInBook";
-import {FineComponent} from "./components/FineComponent";
-import {PayComponent} from "./components/PayComponent";
-
-const appRoutes: Routes = [
-  { path: 'addBorrower', component:  BorrowerComponent },
-  { path:'search', component: SearchComponent},
-  { path: 'addLoan/:isbn', component: CheckoutBook },
-  {path:'checkIn', component:SearchCheckInBook},
-  {path:'addCheckIn/:isbn/:borrowerId', component:CheckInBook},
-  {path:'fine', component: FineComponent},
-  {path:'pay/:cardId',component:PayComponent}
-  /*,
-  {
-    path: 'heroes',
-    component: HeroListComponent,
-    data: { title: 'Heroes List' }
-  },
-  { path: '',
-    redirectTo: '/heroes',
-    pathMatch: 'full'
-  },
-  { path: '**', component: PageNotFoundComponent }*/
-];
-
-
+import {NgModule} from '@angular/core';
+import {BrowserModule} from '@angular/platform-browser';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+import {AppComponent} from './app.component';
+import {RouterModule} from '@angular/router';
+import {HttpModule, JsonpModule} from '@angular/http';
+import {appRoutes} from "./app.routes";
+import {LibraryService} from "./services/library.services";
+import {SearchModule} from "./search/search.module";
+import {OperationModule} from "./operation/operation.module";
+import {UserModule} from "./user/user.module";
 
 @NgModule({
 
@@ -44,9 +18,13 @@ const appRoutes: Routes = [
   	FormsModule,
   	ReactiveFormsModule,
   	HttpModule,
-    JsonpModule
+    JsonpModule,
+    SearchModule,
+    OperationModule,
+    UserModule
   ],
-  declarations: [ AppComponent, BorrowerComponent,  SearchComponent, CheckoutBook, CheckInBook, SearchCheckInBook, FineComponent,PayComponent],
+  declarations: [ AppComponent],
+  providers: [LibraryService],
   bootstrap:    [ AppComponent ]
 })
 export class AppModule { }
