+++ FrontEnd/src/app/model/Author.ts
