+++ FrontEnd/src/app/model/Book.ts
