+++ FrontEnd/src/app/model/CheckInBookModel.ts
