+++ FrontEnd/src/app/operation/fine/fine.component.html
@@ -1,0 +1,43 @@
+<button (click)="addOrUpdateFines()">Update Fines</button>
+<button (click)="showFines()" >Show Fines</button>
+
+<div *ngIf="restResponse" class="form-group">
+  <div [ngSwitch]=restResponse.success>
+    <b *ngSwitchCase="true">{{restResponse.result}}</b>
+    <b *ngSwitchDefault>{{restResponse.error}}</b>
+  </div>
+  <!-- <div *ngIf={{restResponse.success}}>
+  Compilation Result : {{restResponse.result}}
+  <br/>
+  {{restResponse.error}} -->
+</div>
+<!--<div *ngIf="showForm" class="selectedtab">
+ <select [(ngModel)]="selectedValue" (ngModelChange)="onChange($event)">
+   <option selected value="true" (click)="fines()" >Paid</option>
+   <option value="false" (click)="fines()">Not Paid</option>
+ </select>-->
+
+
+<div *ngIf="fineResponse">
+  <!-- <div *ngFor="let f of fineResponse">-->
+  <table border="0">
+    <thead>
+    <tr>
+      <th>Card Id</th>
+      <th>Amount</th>
+    </tr>
+    </thead>
+    <tbody>
+
+    <tr *ngFor="let f of fineResponse">
+      <td>{{f.cardId}}</td>
+      <td>$ {{f.amount}}</td>
+      <td *ngIf="selectedValue=='false'"><button (click)="pay(f.cardId)">Pay</button></td>
+      <td *ngIf="restResponse1">{{restResponse1.result}}</td>
+    </tr>
+
+    </tbody>
+  </table>
+  <!--</div>-->
+
+</div>
