+++ FrontEnd/src/app/operation/pay/pay.component.html
@@ -1,0 +1,25 @@
+<div class="selectedtab">
+  <select [(ngModel)]="selectedValue" (ngModelChange)="onChange($event)">
+    <option selected value="true" >Paid</option>
+    <option value="false">Not Paid</option>
+  </select>
+</div>
+<div *ngIf="fine">
+
+  <table>
+    <tr>
+      <th>Card ID</th>
+      <th>Amount</th>
+      <th>Paid</th>
+    </tr>
+    <tr *ngFor="let f of fine">
+      <td>{{f.fineId}}</td>
+      <td>{{f.fineAmount}}</td>
+      <td>{{f.paid}}</td>
+      <td><button (click)="payFine(f.fineId)">Pay Fine</button></td>
+
+    </tr>
+
+  </table>
+  <div *ngIf="restResponse">{{restResponse.result}}</div>
+</div>
