+++ FrontEnd/src/app/search/check-in/searchCheckInBook.component.ts
@@ -2,20 +2,20 @@
  * Created by asdha on 3/15/2017.
  */
 import {Component} from '@angular/core';
-import {CheckInBookModel} from "../Model/CheckInBookModel";
-import {SearchResult} from "../Model/SearchResult";
-import {LibraryService} from "../services/library.services";
+import {CheckInBookModel} from "../../model/CheckInBookModel";
+import {SearchResult} from "../../model/SearchResult";
+import {LibraryService} from "../../services/library.services";
 import {Router} from "@angular/router";
 
 @Component({
 
-  templateUrl:'app/components/search.checkIn-book.html',
-  styleUrls: [ 'app/components/borrower.component.css' ],
+  templateUrl:'app/search/check-in/searchCheckInBook.component.html',
+  styleUrls: [ 'app/operation/borrower/borrower.component.css' ],
   providers:[LibraryService]
 })
 
 
-export class SearchCheckInBook{
+export class SearchCheckInBookComponent{
 
   checkInBook = new CheckInBookModel("","","");
