+++ FrontEnd/src/app/search/search.module.ts
@@ -1,0 +1,16 @@
+import {NgModule} from "@angular/core";
+import {SearchCheckInBookComponent} from "./check-in/searchCheckInBook.component";
+import {SearchComponent} from "./search.component";
+
+
+@NgModule({
+  imports :[
+
+  ],
+  declarations: [
+    SearchCheckInBookComponent,
+    SearchComponent
+  ]
+})
+
+export class SearchModule{}
