+++ FrontEnd/src/app/user/user.module.ts
@@ -1,0 +1,9 @@
+import {NgModule} from "@angular/core";
+import {CheckInBookComponent} from "./check-in/CheckInBook.component";
+import {CheckOutBookComponent} from "./check-out/checkOutBook.component";
+
+@NgModule({
+  exports: [CheckInBookComponent, CheckOutBookComponent]
+})
+
+export class UserModule {}
